Deleting unnecessary my_dev example
